refactoring(create_table done)
