+++ core/src/server/delivery/request/HasTableRequest.cpp
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "server/grpc_impl/request/CountTableRequest.h"
+#include "server/delivery/request/HasTableRequest.h"
 #include "server/DBWrapper.h"
 #include "utils/Log.h"
 #include "utils/TimeRecorder.h"
@@ -25,21 +25,20 @@
 
 namespace milvus {
 namespace server {
-namespace grpc {
 
-CountTableRequest::CountTableRequest(const std::string& table_name, int64_t& row_count)
-    : GrpcBaseRequest(INFO_REQUEST_GROUP), table_name_(table_name), row_count_(row_count) {
+HasTableRequest::HasTableRequest(const std::string& table_name, bool& has_table)
+    : BaseRequest(INFO_REQUEST_GROUP), table_name_(table_name), has_table_(has_table) {
 }
 
 BaseRequestPtr
-CountTableRequest::Create(const std::string& table_name, int64_t& row_count) {
-    return std::shared_ptr<GrpcBaseRequest>(new CountTableRequest(table_name, row_count));
+HasTableRequest::Create(const std::string& table_name, bool& has_table) {
+    return std::shared_ptr<BaseRequest>(new HasTableRequest(table_name, has_table));
 }
 
 Status
-CountTableRequest::OnExecute() {
+HasTableRequest::OnExecute() {
     try {
-        std::string hdr = "CountTableRequest(table=" + table_name_ + ")";
+        std::string hdr = "HasTableRequest(table=" + table_name_ + ")";
         TimeRecorderAuto rc(hdr);
 
         // step 1: check arguments
@@ -48,18 +47,11 @@ CountTableRequest::OnExecute() {
             return status;
         }
 
-        // step 2: get row count
-        uint64_t row_count = 0;
-        status = DBWrapper::DB()->GetTableRowCount(table_name_, row_count);
+        // step 2: check table existence
+        status = DBWrapper::DB()->HasTable(table_name_, has_table_);
         if (!status.ok()) {
-            if (status.code(), DB_NOT_FOUND) {
-                return Status(SERVER_TABLE_NOT_EXIST, TableNotExistMsg(table_name_));
-            } else {
-                return status;
-            }
+            return status;
         }
-
-        row_count_ = static_cast<int64_t>(row_count);
     } catch (std::exception& ex) {
         return Status(SERVER_UNEXPECTED_ERROR, ex.what());
     }
@@ -67,6 +59,5 @@ CountTableRequest::OnExecute() {
     return Status::OK();
 }
 
-}  // namespace grpc
 }  // namespace server
 }  // namespace milvus
