+++ core/src/server/grpc_impl/GrpcRequestHandler.cpp
@@ -16,24 +16,25 @@
 // under the License.
 
 #include "server/grpc_impl/GrpcRequestHandler.h"
-#include "server/grpc_impl/GrpcRequestScheduler.h"
-#include "server/grpc_impl/request/CmdRequest.h"
-#include "server/grpc_impl/request/CountTableRequest.h"
-#include "server/grpc_impl/request/CreateIndexRequest.h"
-#include "server/grpc_impl/request/CreatePartitionRequest.h"
-#include "server/grpc_impl/request/CreateTableRequest.h"
-#include "server/grpc_impl/request/DeleteByDateRequest.h"
-#include "server/grpc_impl/request/DescribeIndexRequest.h"
-#include "server/grpc_impl/request/DescribeTableRequest.h"
-#include "server/grpc_impl/request/DropIndexRequest.h"
-#include "server/grpc_impl/request/DropPartitionRequest.h"
-#include "server/grpc_impl/request/DropTableRequest.h"
-#include "server/grpc_impl/request/HasTableRequest.h"
-#include "server/grpc_impl/request/InsertRequest.h"
-#include "server/grpc_impl/request/PreloadTableRequest.h"
-#include "server/grpc_impl/request/SearchRequest.h"
-#include "server/grpc_impl/request/ShowPartitionsRequest.h"
-#include "server/grpc_impl/request/ShowTablesRequest.h"
+#include "src/server/delivery/RequestScheduler.h"
+#include "server/delivery/request/CmdRequest.h"
+#include "server/delivery/request/CountTableRequest.h"
+#include "server/delivery/request/CreateIndexRequest.h"
+#include "server/delivery/request/CreatePartitionRequest.h"
+#include "server/delivery/request/CreateTableRequest.h"
+#include "server/delivery/request/DeleteByDateRequest.h"
+#include "server/delivery/request/DescribeIndexRequest.h"
+#include "server/delivery/request/DescribeTableRequest.h"
+#include "server/delivery/request/DropIndexRequest.h"
+#include "server/delivery/request/DropPartitionRequest.h"
+#include "server/delivery/request/DropTableRequest.h"
+#include "server/delivery/request/HasTableRequest.h"
+#include "server/delivery/request/InsertRequest.h"
+#include "server/delivery/request/PreloadTableRequest.h"
+#include "server/delivery/request/SearchRequest.h"
+#include "server/delivery/request/ShowPartitionsRequest.h"
+#include "server/delivery/request/ShowTablesRequest.h"
+#include "src/server/delivery/RequestHandler.h"
 #include "utils/TimeRecorder.h"
 
 #include <vector>
@@ -42,11 +43,62 @@ namespace milvus {
 namespace server {
 namespace grpc {
 
+namespace {
+::milvus::grpc::ErrorCode
+ErrorMap(ErrorCode code) {
+    static const std::map<ErrorCode, ::milvus::grpc::ErrorCode> code_map = {
+        {SERVER_UNEXPECTED_ERROR, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_UNSUPPORTED_ERROR, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_NULL_POINTER, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_INVALID_ARGUMENT, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_FILE_NOT_FOUND, ::milvus::grpc::ErrorCode::FILE_NOT_FOUND},
+        {SERVER_NOT_IMPLEMENT, ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR},
+        {SERVER_CANNOT_CREATE_FOLDER, ::milvus::grpc::ErrorCode::CANNOT_CREATE_FOLDER},
+        {SERVER_CANNOT_CREATE_FILE, ::milvus::grpc::ErrorCode::CANNOT_CREATE_FILE},
+        {SERVER_CANNOT_DELETE_FOLDER, ::milvus::grpc::ErrorCode::CANNOT_DELETE_FOLDER},
+        {SERVER_CANNOT_DELETE_FILE, ::milvus::grpc::ErrorCode::CANNOT_DELETE_FILE},
+        {SERVER_TABLE_NOT_EXIST, ::milvus::grpc::ErrorCode::TABLE_NOT_EXISTS},
+        {SERVER_INVALID_TABLE_NAME, ::milvus::grpc::ErrorCode::ILLEGAL_TABLE_NAME},
+        {SERVER_INVALID_TABLE_DIMENSION, ::milvus::grpc::ErrorCode::ILLEGAL_DIMENSION},
+        {SERVER_INVALID_TIME_RANGE, ::milvus::grpc::ErrorCode::ILLEGAL_RANGE},
+        {SERVER_INVALID_VECTOR_DIMENSION, ::milvus::grpc::ErrorCode::ILLEGAL_DIMENSION},
+
+        {SERVER_INVALID_INDEX_TYPE, ::milvus::grpc::ErrorCode::ILLEGAL_INDEX_TYPE},
+        {SERVER_INVALID_ROWRECORD, ::milvus::grpc::ErrorCode::ILLEGAL_ROWRECORD},
+        {SERVER_INVALID_ROWRECORD_ARRAY, ::milvus::grpc::ErrorCode::ILLEGAL_ROWRECORD},
+        {SERVER_INVALID_TOPK, ::milvus::grpc::ErrorCode::ILLEGAL_TOPK},
+        {SERVER_INVALID_NPROBE, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_INVALID_INDEX_NLIST, ::milvus::grpc::ErrorCode::ILLEGAL_NLIST},
+        {SERVER_INVALID_INDEX_METRIC_TYPE, ::milvus::grpc::ErrorCode::ILLEGAL_METRIC_TYPE},
+        {SERVER_INVALID_INDEX_FILE_SIZE, ::milvus::grpc::ErrorCode::ILLEGAL_ARGUMENT},
+        {SERVER_ILLEGAL_VECTOR_ID, ::milvus::grpc::ErrorCode::ILLEGAL_VECTOR_ID},
+        {SERVER_ILLEGAL_SEARCH_RESULT, ::milvus::grpc::ErrorCode::ILLEGAL_SEARCH_RESULT},
+        {SERVER_CACHE_FULL, ::milvus::grpc::ErrorCode::CACHE_FAILED},
+        {DB_META_TRANSACTION_FAILED, ::milvus::grpc::ErrorCode::META_FAILED},
+        {SERVER_BUILD_INDEX_ERROR, ::milvus::grpc::ErrorCode::BUILD_INDEX_ERROR},
+        {SERVER_OUT_OF_MEMORY, ::milvus::grpc::ErrorCode::OUT_OF_MEMORY},
+    };
+
+    if (code_map.find(code) != code_map.end()) {
+        return code_map.at(code);
+    } else {
+        return ::milvus::grpc::ErrorCode::UNEXPECTED_ERROR;
+    }
+}
+}  // namespace
+
+
 ::grpc::Status
 GrpcRequestHandler::CreateTable(::grpc::ServerContext* context, const ::milvus::grpc::TableSchema* request,
                                 ::milvus::grpc::Status* response) {
-    BaseRequestPtr request_ptr = CreateTableRequest::Create(request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
+    Status status = RequestHandler::CreateTable(
+        request->table_name(),
+        request->dimension(),
+        request->index_file_size(),
+        request->metric_type());
+
+    ConvertToProtoStatus(status, *response);
+
     return ::grpc::Status::OK;
 }
 
@@ -54,9 +106,10 @@ GrpcRequestHandler::CreateTable(::grpc::ServerContext* context, const ::milvus::
 GrpcRequestHandler::HasTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                              ::milvus::grpc::BoolReply* response) {
     bool has_table = false;
-    BaseRequestPtr request_ptr = HasTableRequest::Create(request->table_name(), has_table);
+
+    Status status = RequestHandler::HasTable(request->table_name(), has_table);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_bool_reply(has_table);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
@@ -67,7 +120,7 @@ GrpcRequestHandler::HasTable(::grpc::ServerContext* context, const ::milvus::grp
 GrpcRequestHandler::DropTable(::grpc::ServerContext* context, const ::milvus::grpc::TableName* request,
                               ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = DropTableRequest::Create(request->table_name());
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
+    RequestScheduler::ExecRequest(request_ptr, response);
     return ::grpc::Status::OK;
 }
 
@@ -75,7 +128,7 @@ GrpcRequestHandler::DropTable(::grpc::ServerContext* context, const ::milvus::gr
 GrpcRequestHandler::CreateIndex(::grpc::ServerContext* context, const ::milvus::grpc::IndexParam* request,
                                 ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = CreateIndexRequest::Create(request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
+    RequestScheduler::ExecRequest(request_ptr, response);
     return ::grpc::Status::OK;
 }
 
@@ -84,7 +137,7 @@ GrpcRequestHandler::Insert(::grpc::ServerContext* context, const ::milvus::grpc:
                            ::milvus::grpc::VectorIds* response) {
     BaseRequestPtr request_ptr = InsertRequest::Create(request, response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
@@ -96,7 +149,7 @@ GrpcRequestHandler::Search(::grpc::ServerContext* context, const ::milvus::grpc:
     std::vector<std::string> file_id_array;
     BaseRequestPtr request_ptr = SearchRequest::Create(request, file_id_array, response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_error_code(grpc_status.error_code());
     response->mutable_status()->set_reason(grpc_status.reason());
     return ::grpc::Status::OK;
@@ -113,7 +166,7 @@ GrpcRequestHandler::SearchInFiles(::grpc::ServerContext* context, const ::milvus
     BaseRequestPtr request_ptr =
         SearchRequest::Create(request_mutable->mutable_search_param(), file_id_array, response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_error_code(grpc_status.error_code());
     response->mutable_status()->set_reason(grpc_status.reason());
     return ::grpc::Status::OK;
@@ -124,7 +177,7 @@ GrpcRequestHandler::DescribeTable(::grpc::ServerContext* context, const ::milvus
                                   ::milvus::grpc::TableSchema* response) {
     BaseRequestPtr request_ptr = DescribeTableRequest::Create(request->table_name(), response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_error_code(grpc_status.error_code());
     response->mutable_status()->set_reason(grpc_status.reason());
     return ::grpc::Status::OK;
@@ -136,7 +189,7 @@ GrpcRequestHandler::CountTable(::grpc::ServerContext* context, const ::milvus::g
     int64_t row_count = 0;
     BaseRequestPtr request_ptr = CountTableRequest::Create(request->table_name(), row_count);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_table_row_count(row_count);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
@@ -148,7 +201,7 @@ GrpcRequestHandler::ShowTables(::grpc::ServerContext* context, const ::milvus::g
                                ::milvus::grpc::TableNameList* response) {
     BaseRequestPtr request_ptr = ShowTablesRequest::Create(response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_error_code(grpc_status.error_code());
     response->mutable_status()->set_reason(grpc_status.reason());
     return ::grpc::Status::OK;
@@ -160,7 +213,7 @@ GrpcRequestHandler::Cmd(::grpc::ServerContext* context, const ::milvus::grpc::Co
     std::string result;
     BaseRequestPtr request_ptr = CmdRequest::Create(request->cmd(), result);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_string_reply(result);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
@@ -172,7 +225,7 @@ GrpcRequestHandler::DeleteByDate(::grpc::ServerContext* context, const ::milvus:
                                  ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = DeleteByDateRequest::Create(request);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_error_code(grpc_status.error_code());
     response->set_reason(grpc_status.reason());
     return ::grpc::Status::OK;
@@ -183,7 +236,7 @@ GrpcRequestHandler::PreloadTable(::grpc::ServerContext* context, const ::milvus:
                                  ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = PreloadTableRequest::Create(request->table_name());
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_reason(grpc_status.reason());
     response->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
@@ -194,7 +247,7 @@ GrpcRequestHandler::DescribeIndex(::grpc::ServerContext* context, const ::milvus
                                   ::milvus::grpc::IndexParam* response) {
     BaseRequestPtr request_ptr = DescribeIndexRequest::Create(request->table_name(), response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
@@ -205,7 +258,7 @@ GrpcRequestHandler::DropIndex(::grpc::ServerContext* context, const ::milvus::gr
                               ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = DropIndexRequest::Create(request->table_name());
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_reason(grpc_status.reason());
     response->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
@@ -215,7 +268,7 @@ GrpcRequestHandler::DropIndex(::grpc::ServerContext* context, const ::milvus::gr
 GrpcRequestHandler::CreatePartition(::grpc::ServerContext* context, const ::milvus::grpc::PartitionParam* request,
                                     ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = CreatePartitionRequest::Create(request);
-    GrpcRequestScheduler::ExecRequest(request_ptr, response);
+    RequestScheduler::ExecRequest(request_ptr, response);
     return ::grpc::Status::OK;
 }
 
@@ -224,7 +277,7 @@ GrpcRequestHandler::ShowPartitions(::grpc::ServerContext* context, const ::milvu
                                    ::milvus::grpc::PartitionList* response) {
     BaseRequestPtr request_ptr = ShowPartitionsRequest::Create(request->table_name(), response);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->mutable_status()->set_reason(grpc_status.reason());
     response->mutable_status()->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
@@ -235,10 +288,16 @@ GrpcRequestHandler::DropPartition(::grpc::ServerContext* context, const ::milvus
                                   ::milvus::grpc::Status* response) {
     BaseRequestPtr request_ptr = DropPartitionRequest::Create(request);
     ::milvus::grpc::Status grpc_status;
-    GrpcRequestScheduler::ExecRequest(request_ptr, &grpc_status);
+    RequestScheduler::ExecRequest(request_ptr, &grpc_status);
     response->set_reason(grpc_status.reason());
     response->set_error_code(grpc_status.error_code());
     return ::grpc::Status::OK;
+}
+
+void
+GrpcRequestHandler::ConvertToProtoStatus(const Status& status, ::milvus::grpc::Status& grpc_status) {
+    grpc_status.set_error_code(ErrorMap(status.code()));
+    grpc_status.set_reason(status.message());
 }
 
 }  // namespace grpc
